+++ src/nidmm/metadata/functions_addon.py
@@ -97,9 +97,9 @@ functions_remove_from_python_api = {
 
 # Default values for method parameters
 functions_default_value = {
-    'InitWithOptions':  { 'parameters': { 1: { 'default_value': False, },
-                                          2: { 'default_value': False, },
-                                          3: { 'default_value': '', }, }, },
+    'InitWithOptions':           { 'parameters': { 1: { 'default_value': False, },
+                                                   2: { 'default_value': False, },
+                                                   3: { 'default_value': '""', }, }, },
     'ConfigureMultiPoint':       { 'parameters': { 3: { 'default_value': 'SampleTrigger.IMMEDIATE', },
                                                    4: { 'default_value': -1, }, }, },
     'ConfigureThermocouple':     { 'parameters': { 2: { 'default_value': 'ThermocoupleReferenceJunctionType.FIXED', }, }, },
@@ -107,7 +107,7 @@ functions_default_value = {
     'Fetch':                     { 'parameters': { 1: { 'default_value': -1, }, }, },
     'FetchMultiPoint':           { 'parameters': { 1: { 'default_value': -1, }, }, },
     'FetchWaveform':             { 'parameters': { 1: { 'default_value': -1, }, }, },
-    'GetDevTemp':                { 'parameters': { 1: { 'default_value': '', }, }, },
+    'GetDevTemp':                { 'parameters': { 1: { 'default_value': '""', }, }, },
     'Read':                      { 'parameters': { 1: { 'default_value': -1, }, }, },
     'ReadMultiPoint':            { 'parameters': { 1: { 'default_value': -1, }, }, },
     'ReadWaveform':              { 'parameters': { 1: { 'default_value': -1, }, }, },